Develop

+++ hqet_heavy_to_light/lattice.h
@@ -199,9 +199,9 @@ enum layout_flag_set { HYPER_GRAY_EVENFIRST = 100 } ;
 
 /* Some of these global variables are node dependent */
 /* They are set in "make_lattice()" */
-EXTERN  int	sites_on_node;		/* number of sites on this node */
-EXTERN  int	even_sites_on_node;	/* number of even sites on this node */
-EXTERN  int	odd_sites_on_node;	/* number of odd sites on this node */
+EXTERN  size_t	sites_on_node;		/* number of sites on this node */
+EXTERN  size_t	even_sites_on_node;	/* number of even sites on this node */
+EXTERN  size_t	odd_sites_on_node;	/* number of odd sites on this node */
 EXTERN  int	number_of_nodes;	/* number of nodes in use */
 EXTERN  int  this_node;		/* node number of this node */
 EXTERN Real *wall_template; /* pointer to array of wall heights */
