+++ schroed_ks_dyn/setup.c
@@ -85,7 +85,7 @@ int prompt=0,status;
     nflavors=par_buf.nflavors;
 
     number_of_nodes = numnodes();
-    volume=nx*ny*nz*nt;
+    volume=(size_t)nx*ny*nz*nt;
     total_iters=0;
     return(prompt);
 }
