+++ wilson_dynamical/setup_qdp.c
@@ -88,7 +88,7 @@ int prompt,status;
 
     this_node = mynode();
     number_of_nodes = numnodes();
-    volume=nx*ny*nz*nt;
+    volume=(size_t)nx*ny*nz*nt;
     total_iters=0;
     return(prompt);
 }
